backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
Pillow==12.1.1
python-dotenv==1.0.1
requests==2.33.1
pymongo==4.10.1
google-auth==2.38.0
